feat: calling tcp connect from http connect, chore: moved connect functions away from spec
