+++ src/pkg/server/tcp.rs
@@ -1,4 +1,5 @@
 use rand::Rng;
+use tokio::net::TcpStream;
 use tokio::task::JoinSet;
 use tokio::{
     io::{AsyncReadExt, AsyncWriteExt},
@@ -12,6 +13,17 @@ use crate::{
 use async_trait::async_trait;
 
 use super::{proxy::spawn_tcp_server, ForwardRoutes, SpawnServers};
+
+
+impl TcpRoute {
+    pub async fn connect(&self) -> TcpStream {
+        let destination = format!("{}:{}", &self.target_host, &self.target_port);
+        tracing::debug!("connecting to remote: {}", &destination);
+        let conn = TcpStream::connect(&destination).await.unwrap();
+        tracing::info!("✅ Connected to upstream: {:?}", &self);
+        conn
+    }
+}
 
 #[async_trait]
 impl SpawnServers for TcpRoutes {
